Extract native repository worker lifecycle and relocate smoke tests
